修改数据表id不为long型的bug 1、修改数据表类型 2、逆向工程重新生成 3、swagger接口重新生成 4、业务代码修改参数类型

+++ src/main/java/org/garen/mc/mybatis/mapper/CarouselMapper.java
@@ -11,7 +11,7 @@ public interface CarouselMapper <T,Q,PK extends Serializable> extends CommonMapp
 
     int deleteByExample(CarouselExample example);
 
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(Long id);
 
     int insert(Carousel record);
 
@@ -19,7 +19,7 @@ public interface CarouselMapper <T,Q,PK extends Serializable> extends CommonMapp
 
     List<Carousel> selectByExample(CarouselExample example);
 
-    Carousel selectByPrimaryKey(Integer id);
+    Carousel selectByPrimaryKey(Long id);
 
     int updateByExampleSelective(@Param("record") Carousel record, @Param("example") CarouselExample example);
